迁移活动脚本
迁移活动脚本
